fix CGLSetParameter argument conversion 10.5 build error

+++ libs/openFrameworks/app/ofAppRunner.cpp
@@ -323,7 +323,7 @@ void ofSetVerticalSync(bool bSync){
 	//--------------------------------------
 	#ifdef TARGET_OSX
 	//--------------------------------------
-		long sync = bSync == true ? 1 : 0;
+		GLint sync = bSync == true ? 1 : 0;
 		CGLSetParameter (CGLGetCurrentContext(), kCGLCPSwapInterval, &sync);
 	//--------------------------------------
 	#endif
